requirements.txt
numpy>=1.11.0
scipy>=0.12.0
matplotlib>=1.5.0
pillow


# for Google Vision API
google-cloud-vision
google-cloud
python-rapidjson
google_images_download
